==> notebooks/1.0-jc-selection.ipynb <==
### Importar as bibliotecas

%run ../services/imports.py
%run ../services/visualization.py
import services.list_columns
import services.alterarColuna as atc

## Dados de Homologação

dados_homologados_2016 = pd.read_excel("../dados/raw/POSCOMP2016-2017/2016/homologados_dados_contatos_poscomp_2016.xlsx")
dados_homologados_2017 = pd.read_excel("../dados/raw/POSCOMP2016-2017/2017/homologados_dados_contatos_poscomp_2017.xlsx")
dados_homologados_2018 = pd.read_excel("../dados/raw/POSCOMP2018-2019/2018/homologados_dados_contatos_poscomp_2018.xlsx")
dados_homologados_2019 = pd.read_excel("../dados/raw/POSCOMP2018-2019/2019/homologados_dados_contatos_poscomp_2019.xlsx")

# dados_homologados_2016 = dados_homologados_2016.assign(ANO='2016')
# dados_homologados_2017 = dados_homologados_2017.assign(ANO='2017')
# dados_homologados_2018 = dados_homologados_2018.assign(ANO='2018')
# dados_homologados_2019 = dados_homologados_2019.assign(ANO='2019')

dados_homologados = pd.concat([dados_homologados_2016, dados_homologados_2017, dados_homologados_2018, dados_homologados_2019], axis=0)
dados_homologados = dados_homologados.drop(columns=['TIPO DE DOCUMENTO', 'DOCUMENTO','NOME DO PAI', 
                                                    'NOME DA MÃE', 'EMAIL', 'DDD', 'FONE', 'DDD CEL', 'CELULAR', 'SENHA', 'DATA INSCRIÇÃO',
                                                    'DEFICIENTE', 'NECESSIDADES ESPECIAIS', 'DESCRICAO DA NECESSIDADE', 'ESTRANGEIRO', 
                                                    'CARGO','DS CARGO', 'COMPL.', 'ENDERECO.', 'BAIRRO', 'CEP', 'NUMERO', 'LOCAL DE PROVA', 'CIDADE', 'CÓDIGO'])

dados_homologados

## Notas das respostas

#### Base de 2016 a 2019

dados_notasRespostas_2016 = pd.read_excel("../dados/raw/POSCOMP2016-2017/2016/Notas_poscomp_2016.xlsx")
dados_notasRespostas_2017 = pd.read_excel("../dados/raw/POSCOMP2016-2017/2017/Notas_poscomp_2017.xlsx")
dados_notasRespostas_2018 = pd.read_excel("../dados/raw/POSCOMP2018-2019/2018/Notas_poscomp_2018.xlsx")
dados_notasRespostas_2019 = pd.read_excel("../dados/raw/POSCOMP2018-2019/2019/Notas_poscomp_2019.xlsx")

dados_notasRespostas_2016.drop(columns=['NOTA TOTAL Prova Discursiva',' OBSERVAÇÃO Prova Discursiva'], axis=1, inplace=True)
dados_notasRespostas_2018.drop(columns=['NOTA TOTAL Prova Discursiva',' OBSERVAÇÃO Prova Discursiva'], axis=1, inplace=True)

dados_notasRespostas_2016 = dados_notasRespostas_2016.assign(ANO='2016')
dados_notasRespostas_2017 = dados_notasRespostas_2017.assign(ANO='2017')
dados_notasRespostas_2018 = dados_notasRespostas_2018.assign(ANO='2018')
dados_notasRespostas_2019 = dados_notasRespostas_2019.assign(ANO='2019')

dados_notasRespostas_2016 = dados_notasRespostas_2016[services.list_columns.lista_2016]
dados_notasRespostas_2017 = dados_notasRespostas_2017[services.list_columns.lista_2017]
dados_notasRespostas_2018 = dados_notasRespostas_2018[services.list_columns.lista_2018]
dados_notasRespostas_2019 = dados_notasRespostas_2019[services.list_columns.lista_2019]

dados_notasRespostas_2016.shape, dados_notasRespostas_2017.shape, dados_notasRespostas_2019.shape, dados_notasRespostas_2019.shape

dados_notasRespostas_2016 = dados_notasRespostas_2016.rename(columns=atc.alteraColunas_2016)
dados_notasRespostas_2017 = dados_notasRespostas_2017.rename(columns=atc.alteraColunas_2017)
dados_notasRespostas_2018 = dados_notasRespostas_2018.rename(columns=atc.alteraColunas_2018)
dados_notasRespostas_2019 = dados_notasRespostas_2019.rename(columns=atc.alteraColunas_2019)

dados_notasRespostas = pd.concat([dados_notasRespostas_2016, dados_notasRespostas_2017, dados_notasRespostas_2018, dados_notasRespostas_2019], axis=0)

dados_notasRespostas.drop_duplicates(subset='INSCRICAO', inplace=True)

## Mesclar os dados

dados_homologados.shape

dados_notasRespostas.shape

data_poscomp = dados_homologados.merge(dados_notasRespostas,
                        how='inner',
                        left_on=['INSCRICAO'],
                        right_on=['INSCRICAO'])

data_poscomp.shape

## Column transformation 

data_poscomp.columns = data_poscomp.columns.str.lower()

data_poscomp = data_poscomp.applymap(lambda x: x.lower() if isinstance(x, str) else x)
data_poscomp

data_poscomp.rename(columns= {'data nasc.': 'idade', 'presente dia1': 'presença'}, inplace=True)

data_poscomp.columns

data_poscomp.columns = [column.replace(" ", "_") for column in (data_poscomp.columns)]
data_poscomp.columns = [column.replace(",", "") for column in (data_poscomp.columns)]

data_poscomp

## Salvar dados

data_poscomp.to_csv('../dados/processed/2016-2019/dataset_poscomp.csv', index=False)

# dados_notasRespostas.to_csv('dados/notas_poscomp/dados_notasRespostas.csv')
