==> notebooks/1.1-jc-selection.ipynb <==
# Working with the new POSCOMP data (2022-2023)

## Import libraries

%run ../services/imports.py
%run ../services/visualization.py

dataset_2022 = pd.read_excel('../dados/raw/POSCOMP_2022.xlsx')
dataset_2023 = pd.read_excel('../dados/raw/POSCOMP_2023.xlsx')

dataset_2022.columns = dataset_2022.columns.str.lower()
dataset_2023.columns = dataset_2023.columns.str.lower()

dataset_2022.columns = [column.replace(" ", "_") for column in (dataset_2022.columns)]
dataset_2022.columns = [column.replace(",", "") for column in (dataset_2022.columns)]

dataset_2023.columns = [column.replace(" ", "_") for column in (dataset_2023.columns)]
dataset_2023.columns = [column.replace(",", "") for column in (dataset_2023.columns)]

dataset_2023

dataset_2022.drop(columns=['cód._cargo', 'cidade_local_prova','local_da_prova'], inplace=True)
dataset_2023.drop(columns=['cód._cargo', 'cidade_candidato', 'cód._especialidade'], inplace=True)

dataset_2022 = dataset_2022.applymap(lambda x: x.lower() if isinstance(x, str) else x)
dataset_2023 = dataset_2023.applymap(lambda x: x.lower() if isinstance(x, str) else x)

def process_dataframe(df, column_to_drop):
    """
    Processa o DataFrame separando uma coluna específica e removendo duplicatas.

    Parâmetros:
    df (DataFrame): O DataFrame original.
    column_to_drop (str): O nome da coluna a ser separada e removida.

    Retorna:
    df_x (DataFrame): DataFrame contendo apenas a coluna separada.
    df_y (DataFrame): DataFrame sem a coluna separada, com duplicatas removidas e índice adicionado.
    """
    df_x = df[[column_to_drop]].copy()
    df_y = df.drop(column_to_drop, axis=1).drop_duplicates().reset_index(drop=True)
    df_x['index_col'] = df_x.index
    df_y['index_col'] = df_y.index
    
    return df_x, df_y

# Exemplo de uso:
# jean_x, jean_y = process_dataframe(jean, 'especialidade')


df_x, df_y = process_dataframe(dataset_2022, 'especialidade')
dataframe_complete_2022 = df_y.merge(df_x,
                                  how='inner',
                                  left_on=['index_col'],
                                  right_on=['index_col'])
dataframe_complete_2022

df_x, df_y = process_dataframe(dataset_2023, 'especialidade')
dataframe_complete_2023 = df_y.merge(df_x,
                                  how='inner',
                                  left_on=['index_col'],
                                  right_on=['index_col'])
dataframe_complete_2023

dataframe_complete_2022.rename(columns=lambda x: x.replace('nota_', ''), inplace=True)

dataframe_complete_2022 = dataframe_complete_2022.drop(['necessidades_especiais','index_col','total_to'], axis=1)

dataframe_complete_2022.columns

dataframe_complete_2022.rename(columns={'data_de_nascimento':'idade', 'presente_dia1':'presença'}, 
                               inplace=True)

list_order_columns = ['idade','sexo','estado','presença','cargo','especialidade','algoritmos_e_estrutura_de_dados',
 'análise_combinatória',
 'análise_de_algoritmos',
 'arquitetura_e_organização_de_computadores',
 'banco_de_dados',
 'calculo_diferencial_e_integral',
 'circuitos_digitais',
 'compiladores',
 'computação_gráfica',
 'engenharia_de_software',
 'geometria_analítica',
 'inteligência_artificial',
 'linguagens_de_programação',
 'linguagens_formais_autômatos_e_computabilidade',
 'lógica_matemática',
 'matemática_discreta',
 'organização_de_arquivos_e_dados',
 'probabilidade_e_estatística',
 'processamento_de_imagens',
 'redes_de_computadores',
 'sistemas_distribuídos',
 'sistemas_operacionais',
 'teoria_dos_grafos',
 'técnicas_de_programação',
 'álgebra_linear']

dataframe_complete_2022 = dataframe_complete_2022.reindex(labels=list_order_columns, axis=1)

dataframe_complete_2022

dataframe_complete_2022 = dataframe_complete_2022.assign(ano='2022')

dataframe_complete_2022.to_csv('../dados/processed/2016-2019/dataset_poscomp2022.csv', index=False)

### Trabalhando com o dados do POSCOMP 2023

dataframe_complete_2023

dataframe_complete_2023.rename(columns=lambda x: x.replace('nota_', ''), inplace=True)

dataframe_complete_2023 = dataframe_complete_2023.drop(['nec_esp','index_col','total_to'], axis=1)

dataframe_complete_2023.rename(columns={'data_de_nascimento':'idade', 'presente_dia1':'presença',
                                       'estado_candidato': 'estado'}, inplace=True)

dataframe_complete_2023 = dataframe_complete_2023.reindex(labels=['idade', 'sexo', 'estado', 'presença','cargo','especialidade',
                                       'matemática','fundamentos_da_computação', 'tecnologia_de_computadores'], axis=1)

dataframe_complete_2023 = dataframe_complete_2023.assign(ano='2023')

dataframe_complete_2023

dataframe_complete_2023.to_csv('../dados/processed/2016-2019/dataset_poscomp2023.csv', index=False)

# desired_columns = list(dataset_2022.columns)
# desired_columns.remove("especialidade")
# desired_columns

# # Assuming dataset_2022 is your DataFrame with 60 columns and "especialidade" column
# # Create a list to store individual DataFrames
# dfs = []

# # Group by all columns except "especialidade" and iterate over groups
# for _, group in dataset_2022.groupby(list(dataset_2022.columns.difference(['especialidade']))):
#     # Get unique values of "especialidade" for each group
#     especialidades = group['especialidade'].unique()
    
#     # Create a separate DataFrame for each unique "especialidade" value
#     for especialidade in especialidades:
#         # Filter rows based on "especialidade" value
#         filtered_group = group[group['especialidade'] == especialidade].drop(columns=['especialidade'])
        
#         # Append the filtered DataFrame to the list
#         dfs.append(filtered_group)

# # Displaying the first few rows of each separate DataFrame
# for i, df_individual in enumerate(dfs):
#     print(f"DataFrame {i+1}:")
#     print(df_individual.head())
#     print()


# students = []

# for i, row in dataset_2022_dupli_present.iterrows():
#     for column in list(dataset_2022.columns):
#         parts = []
#         if column != "especialidade":
#             if dataset_2022.dtypes[column] == "object":
#                 part = f"({column} == '{row[column]}')"
#             else:
#                 part = f"({column} == {row[column]})"
#             print(part)
#             parts.append(part)

#     query = " & ".join(parts)
#     print(query)
                
#     current_df = dataset_2022.query(query)
#     display(current_df)
#     break
#     # students.append(current_df)

# # students[0]
