==> notebooks/2.0-jc-organizing_dataframe.ipynb <==
# Organização dos dados

<p>Neste notebook serão organizados os dados dos participantes que realizaram o POSCOMP dos anos de 2016 a 2022. Os participantes dos anos de 2023 não entram porque a SBC determinou que não serão contatos os temas aplicados nas provas, apenas os eixos centrais, os quais são: Matemática, Fundamentos da Computação e Tecnologia da Computação.</p>

#### Importação das bibliotecas

%run ../services/imports.py
%run ../services/visualization.py

dataframe2016_2019 = pd.read_csv("../dados/processed/2016-2019/dataset_poscomp.csv")

dataframe2022 = pd.read_csv("../dados/processed/2016-2019/dataset_poscomp2022.csv")

dataframe2016_2019.drop(columns=['cpf','nome','inscricao'], inplace=True)

list_order_columns = ['idade','sexo','estado','presença','cargo','especialidade','algoritmos_e_estruturas_de_dados',
 'análise_combinatória',
 'análise_de_algoritmos',
 'arquitetura_e_organização_de_computadores',
 'banco_de_dados',
 'cálculo_diferencial_e_integral',
 'circuitos_digitais',
 'compiladores',
 'computação_gráfica',
 'engenharia_de_software',
 'geometria_analítica',
 'inteligência_artificial',
 'linguagens_de_programação',
 'linguagens_formais_autômatos_e_computabilidade',
 'lógica_matemática',
 'matemática_discreta',
 'organização_de_arquivos_e_dados',
 'probabilidade_e_estatística',
 'processamento_de_imagens',
 'redes_de_computadores',
 'sistemas_distribuídos',
 'sistemas_operacionais',
 'teoria_dos_grafos',
 'técnicas_de_programação',
 'álgebra_linear', 'ano']

dataframe2016_2019.columns

dataframe2016_2019 = dataframe2016_2019.reindex(list_order_columns, axis=1)

dataframe2016_2019.info()

list(dataframe2022.columns)

rename_columns = {
    'algoritmos_e_estrutura_de_dados': 'algoritmos_e_estruturas_de_dados',
    'álgebra_linear': 'algebra_linear',
    'análise_combinatória': 'analise_combinatoria',
    'análise_de_algoritmos': 'analise_de_algoritmos',
    'arquitetura_e_organização_de_computadores': 'arquitetura_e_organizacao_de_computadores',
    'banco_de_dados': 'banco_de_dados',
    'cálculo_diferencial_e_integral': 'calculo_diferencial_e_integral',
    'circuitos_digitais': 'circuitos_digitais',
    'compiladores': 'compiladores',
    'computação_gráfica': 'computacao_grafica',
    'engenharia_de_software': 'engenharia_de_software',
    'geometria_analítica': 'geometria_analitica',
    'inteligência_artificial': 'inteligencia_artificial',
    'linguagens_de_programação': 'linguagens_de_programacao',
    'linguagens_formais_autômatos_e_computabilidade': 'linguagens_formais_automatos_e_computabilidade',
    'lógica_matemática': 'logica_matematica',
    'matemática_discreta': 'matematica_discreta',
    'organização_de_arquivos_e_dados': 'organizacao_de_arquivos_e_dados',
    'probabilidade_e_estatística': 'probabilidade_e_estatistica',
    'processamento_de_imagens': 'processamento_de_imagens',
    'redes_de_computadores': 'redes_de_computadores',
    'sistemas_distribuídos': 'sistemas_distribuidos',
    'sistemas_operacionais': 'sistemas_operacionais',
    'técnicas_de_programação': 'tecnicas_de_programacao',
    'teoria_dos_grafos': 'teoria_dos_grafos',
    'ANO': 'ano'
}

rename_columns2 = {
    'algoritmos_e_estrutura_de_dados': 'algoritmos_e_estrutura_de_dados',
    'álgebra_linear': 'algebra_linear',
    'análise_combinatória': 'analise_combinatoria',
    'análise_de_algoritmos': 'analise_de_algoritmos',
    'arquitetura_e_organização_de_computadores': 'arquitetura_e_organizacao_de_computadores',
    'banco_de_dados': 'banco_de_dados',
    'cálculo_diferencial_e_integral': 'calculo_diferencial_e_integral',
    'circuitos_digitais': 'circuitos_digitais',
    'compiladores': 'compiladores',
    'computação_gráfica': 'computacao_grafica',
    'engenharia_de_software': 'engenharia_de_software',
    'geometria_analítica': 'geometria_analitica',
    'inteligência_artificial': 'inteligencia_artificial',
    'linguagens_de_programação': 'linguagens_de_programacao',
    'linguagens_formais_autômatos_e_computabilidade': 'linguagens_formais_automatos_e_computabilidade',
    'lógica_matemática': 'logica_matematica',
    'matemática_discreta': 'matematica_discreta',
    'organização_de_arquivos_e_dados': 'organizacao_de_arquivos_e_dados',
    'probabilidade_e_estatística': 'probabilidade_e_estatistica',
    'processamento_de_imagens': 'processamento_de_imagens',
    'redes_de_computadores': 'redes_de_computadores',
    'sistemas_distribuídos': 'sistemas_distribuidos',
    'sistemas_operacionais': 'sistemas_operacionais',
    'técnicas_de_programação': 'tecnicas_de_programacao',
    'teoria_dos_grafos': 'teoria_dos_grafos',
    'ANO': 'ano'
}

dataframe2022.rename(columns=rename_columns, inplace=True)

dataframe2016_2019.head()

dataframe2016_2019.rename(columns=rename_columns, inplace=True)

dataframe2022.columns

#### Concatenação

dataframe_completo = pd.concat([dataframe2016_2019, dataframe2022])

dataframe_completo.info()

dataframe_completo['sexo'].replace({'m':'masculino', 'f':'feminino'}, inplace=True)

### Organizar o dataframe dos participantes de 2023

<p> Nesta parte serão trabalhados com os dados do POSCOMP 2023. Nestes dados não estão presentes todos os temas das disciplinas, apenas o eixos de Matemática, Fundamentos da Computação e Tecnologia da Computação. </p>

dataframe_2023 = pd.read_csv('../dados/processed/2016-2019/dataset_poscomp2023.csv')

dataframe_2023.rename(columns={'matemática':'matematica'}, inplace=True)

dataframe_2023

## Remover os temas da prova

dataframe_completo["matematica"] = dataframe_completo.loc[
    :,
    [
        "matematica_discreta",
        "calculo_diferencial_e_integral",
        "algebra_linear",
        "geometria_analitica",
        "analise_combinatoria",
        "probabilidade_e_estatistica",
        "logica_matematica",
    ],
].sum(axis=1)

dataframe_completo["fund_computacao"] = dataframe_completo.loc[
    :,
    [
        "organizacao_de_arquivos_e_dados",
        "tecnicas_de_programacao",
        "circuitos_digitais",
        "analise_de_algoritmos",
        "teoria_dos_grafos",
        "linguagens_formais_automatos_e_computabilidade",
        "algoritmos_e_estruturas_de_dados",
        "linguagens_de_programacao",
        "sistemas_operacionais",
        "arquitetura_e_organizacao_de_computadores",
    ],
].sum(axis=1)

dataframe_completo["tec_computacao"] = dataframe_completo.loc[
    :,
    [
        "redes_de_computadores",
        "computacao_grafica",
        "inteligencia_artificial",
        "engenharia_de_software",
        "sistemas_distribuidos",
        "compiladores",
        "processamento_de_imagens",
        "banco_de_dados",
    ],
].sum(axis=1)

dataframe_completo = dataframe_completo.query('estado != "ou"')

dataframe_completo.to_csv('../dados/processed/df_poscomp2016-2022.csv', index=False)

dataframe_completo = dataframe_completo.loc[:, ['idade','sexo','estado','presença','cargo','especialidade','matematica','fund_computacao','tec_computacao','ano']]

dataframe_2023.rename(columns={'fundamentos_da_computação':'fund_computacao', 
                       'tecnologia_de_computadores':'tec_computacao'}, inplace=True)

dataframe_2023 = dataframe_2023.query('estado != "ou"')

dados_poscomp = pd.concat([dataframe_completo, dataframe_2023])

dados_poscomp.to_csv('../dados/processed/dados_poscomp_completo.csv', index=False)
