==> notebooks/4.0-jc-feature_engineering.ipynb <==
## Importação das bibliotecas

%run ../services/imports.py
%run ../services/visualization.py
from services.function import Function

df_poscomp = pd.read_csv("../dados/processed/df_poscomp2016-2022.csv")

# converter os type das colunas
df_poscomp["ano"] = df_poscomp["ano"].astype("object")
# Primeiramente, será excluídos candidatos que realizaram o exame no Peru
df_poscomp = df_poscomp.query("estado != 'ou' & presença != 'n'")
df_poscomp["idade"] = df_poscomp.apply(Function.calculate_age, axis=1)
df_poscomp = df_poscomp.query('idade > 0')

df_poscomp.dropna(inplace=True)

# df_poscomp.drop(columns=['presença','cargo'], inplace=True)

# identificar os atributos relacionados

df_poscomp.isnull().sum()

def encode_categorical_columns(df, columns):
    le = LabelEncoder()
    for col in columns:
        df[col] = le.fit_transform(df[col])
    return df

categorical_columns = ['sexo','estado','especialidade']
X_dataset = encode_categorical_columns(df_poscomp, categorical_columns)
